requirements.txt
flask==3.0.3
feedparser==6.0.11
trafilatura==1.12.2
requests==2.32.3
python-dateutil==2.9.0.post0
markdown==3.7
pyyaml==6.0.2
